Aayushi_Harrypotter_contribution.py
On entering your name you will be sorted into a house, given a character you are similar to along with an attribute. Then an anigmus, favorite subject and  the professors whose favorite you'd be. This is all with respect to the harry potter series and the answers you get and randomly assigned. Its a simple and fun game to play.